requirements.txt
streamlit==1.12.0
nltk==3.7
pdfplumber==0.7.4
python-docx==0.8.11
spacy==3.2.4
